clean and rename the files
